+++ ios/DrumPicker.mm
@@ -1,0 +1,7 @@
+#import <React/RCTBridgeModule.h>
+#import <React/RCTViewManager.h>
+
+// Fabric registration: `DrumPickerViewCls` in DrumPickerView.mm + codegen `componentProvider`.
+
+@interface RCT_EXTERN_MODULE(DrumPickerViewManager, RCTViewManager)
+@end